+++ DistributedMutex.hpp
@@ -3,6 +3,7 @@
 
 #ifdef __WITH_MPI
 
+#include <cstdint>
 #include <mpi.h>
 #include <memory>
 #include <mpi_utils/mpi_commands.hpp>
@@ -12,7 +13,7 @@ class DistributedMutex
 {
 public:
     DistributedMutex(const MPI_Comm &comm, rank_t rank, RDMA_Type rdma_type);
-    
+
     ~DistributedMutex();
 
     void Lock(void);
@@ -26,7 +27,7 @@ public:
 private:
     MPI_Comm comm;
     rank_t rank;
-    std::unique_ptr<RemoteMemoryAgent<int>> agent;
+    std::unique_ptr<RemoteMemoryAgent<uint64_t>> agent;
     bool destroyed;
 };
